12.15

+++ Assets/Homeworks/homework 7/MinMax.cs
@@ -5,14 +5,15 @@ using UnityEngine;
 public class MinMax : MonoBehaviour
 {
     [SerializeField] float [] array;
-    [SerializeField] int min, max;
+    [SerializeField] float min, max;
     
 
     private void OnValidate()
     {
+
         for (int i = 0; i < array.Length; i++)
         {
-            int random = Random.Range(0, 100);
+            float random = Random.Range(0, 100);
             array[i] = random;
 
         }
